Change DB chat_id type to BigInteger

+++ app/models.py
@@ -1,6 +1,6 @@
 """SQLAlchemy ORM 模型:三张表。"""
 
-from sqlalchemy import Column, DateTime, Integer, String, Text, func
+from sqlalchemy import BigInteger, Column, DateTime, Integer, String, Text, func
 from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
@@ -12,7 +12,7 @@ class Message(Base):
     __tablename__ = "messages"
 
     id = Column(Integer, primary_key=True)
-    chat_id = Column(Integer, index=True)
+    chat_id = Column(BigInteger, index=True)
     role = Column(String(20))  # "user" | "assistant"
     content = Column(Text)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
@@ -24,7 +24,7 @@ class Thought(Base):
     __tablename__ = "thoughts"
 
     id = Column(Integer, primary_key=True)
-    chat_id = Column(Integer, index=True)
+    chat_id = Column(BigInteger, index=True)
     content = Column(Text)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
 
@@ -35,7 +35,7 @@ class ScheduledWakeup(Base):
     __tablename__ = "scheduled_wakeups"
 
     id = Column(Integer, primary_key=True)
-    chat_id = Column(Integer, index=True)
+    chat_id = Column(BigInteger, index=True)
     fire_at = Column(DateTime(timezone=True), index=True)
     reason = Column(Text)
     status = Column(String(20), default="pending")  # pending | fired | cancelled
